+++ src/main/java/com/example/forumproject/models/dtos/FullProfileUserDto.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class FullProfileUserDto {
 
-    private int id;
+    private Long id;
     private String first_name;
     private String last_name;
     private String username;
@@ -17,7 +17,7 @@ public class FullProfileUserDto {
     private List<PostOutDto> user_posts;
 
 
-    public FullProfileUserDto(int id, String first_name, String last_name, String username, String email, String account_type, String account_status, String profile_photo) {
+    public FullProfileUserDto(Long id, String first_name, String last_name, String username, String email, String account_type, String account_status, String profile_photo) {
         this.id = id;
         this.first_name = first_name;
         this.last_name = last_name;
@@ -28,11 +28,11 @@ public class FullProfileUserDto {
         this.profile_photo = profile_photo;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
